requirements.txt
Flask==2.3.2
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
plotly==5.14.1
gunicorn==20.1.0
pytz==2023.3
psycopg2-binary==2.9.3
python-dotenv==0.19.1